refactor: reduce verbosity in render dispatch and layout rendering
- split the layout-aware string rendering out of default.py into a dedicated
  floss/render/layout.py module, so default.py only orchestrates the classic
  and recovered-string sections (was ~860 lines mixing two unrelated styles)
- collapse the three recovered-string section blocks into one loop over a
  (name, enabled, strings, renderer) table
- extract render_text() in main.py to share the plain/summary/default render
  dispatch between the results-load and analysis paths
- generate the six --section/--structure/--tag include/exclude arguments from
  a single table in cli.py
- share language_value() between render_meta and the summary renderer
